# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangeDetailResponse": {
        "properties": {
          "author": {
            "title": "Author",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "diff": {
            "title": "Diff",
            "type": "string"
          },
          "diff_summary": {
            "$ref": "#/components/schemas/DiffSummaryResponse"
          },
          "diff_truncated": {
            "title": "Diff Truncated",
            "type": "boolean"
          },
          "findings_summary": {
            "$ref": "#/components/schemas/FindingsSummaryResponse"
          },
          "head_sha": {
            "title": "Head Sha",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ChangeKind"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "review_status": {
            "$ref": "#/components/schemas/ChangeReviewStatus"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            },
            "title": "Reviews",
            "type": "array"
          },
          "run": {
            "title": "Run",
            "type": "integer"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/ChangeStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "kind",
          "ref",
          "head_sha",
          "title",
          "author",
          "url",
          "diff_truncated",
          "status",
          "review_status",
          "run",
          "created_at",
          "diff_summary",
          "stale",
          "diff",
          "reviews",
          "findings_summary"
        ],
        "title": "ChangeDetailResponse",
        "type": "object"
      },
      "ChangeEventResponse": {
        "description": "Evento de la l\u00ednea de tiempo con el payload reducido: solo `agent` y `review_id`, nunca la\nsalida cruda ni el texto de un error.",
        "properties": {
          "agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "review_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Id"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "created_at",
          "agent",
          "review_id"
        ],
        "title": "ChangeEventResponse",
        "type": "object"
      },
      "ChangeKind": {
        "enum": [
          "commit",
          "pr"
        ],
        "title": "ChangeKind",
        "type": "string"
      },
      "ChangePageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChangeSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "ChangePageResponse",
        "type": "object"
      },
      "ChangeReviewStatus": {
        "enum": [
          "pending",
          "running",
          "partial_failed",
          "failed",
          "completed"
        ],
        "title": "ChangeReviewStatus",
        "type": "string"
      },
      "ChangeStatus": {
        "enum": [
          "pending"
        ],
        "title": "ChangeStatus",
        "type": "string"
      },
...
      "DiffFileResponse": {
        "properties": {
          "additions": {
            "title": "Additions",
            "type": "integer"
          },
          "deletions": {
            "title": "Deletions",
            "type": "integer"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "additions",
          "deletions"
        ],
        "title": "DiffFileResponse",
        "type": "object"
      },
      "DiffSummaryResponse": {
        "description": "Archivos y l\u00edneas del diff, calculados al ingerir. `files` se acota a 200 entradas;\n`files_changed` es el recuento real.",
        "properties": {
          "additions": {
            "title": "Additions",
            "type": "integer"
          },
          "deletions": {
            "title": "Deletions",
            "type": "integer"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/DiffFileResponse"
            },
            "title": "Files",
            "type": "array"
          },
          "files_changed": {
            "title": "Files Changed",
            "type": "integer"
          }
        },
        "required": [
          "files_changed",
          "additions",
          "deletions",
          "files"
        ],
        "title": "DiffSummaryResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FindingsSummaryResponse": {
        "description": "Findings del run actual con las severidades normalizadas (`other`: no reconocidas).",
        "properties": {
          "by_severity": {
            "$ref": "#/components/schemas/SeverityCountsResponse"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "by_severity"
        ],
        "title": "FindingsSummaryResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestCommitRequest": {
        "additionalProperties": false,
        "properties": {
          "author": {
            "default": "",
            "title": "Author",
            "type": "string"
          },
          "diff": {
            "default": "",
            "title": "Diff",
            "type": "string"
          },
          "head_sha": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Head Sha",
            "type": "string"
          },
          "project": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Project",
            "type": "string"
          },
          "ref": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Ref",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "url": {
            "default": "",
            "maxLength": 1000,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "project",
          "ref",
          "head_sha"
        ],
        "title": "IngestCommitRequest",
        "type": "object"
      },
      "IngestCommitResponse": {
        "properties": {
          "change_id": {
            "format": "uuid",
            "title": "Change Id",
            "type": "string"
          },
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "diff_truncated": {
            "title": "Diff Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "change_id",
          "diff_truncated",
          "created"
        ],
        "title": "IngestCommitResponse",
        "type": "object"
      },
      "IngestPrRequest": {
        "additionalProperties": false,
        "description": "Mismos campos y l\u00edmites que un commit: `ref` es la rama o `refs/pull/N/head` y\n`head_sha` el \u00faltimo commit de la cabeza del PR.",
        "properties": {
          "author": {
            "default": "",
            "title": "Author",
            "type": "string"
          },
          "diff": {
            "default": "",
            "title": "Diff",
            "type": "string"
          },
          "head_sha": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Head Sha",
            "type": "string"
          },
          "project": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Project",
            "type": "string"
          },
          "ref": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Ref",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "url": {
            "default": "",
            "maxLength": 1000,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "project",
          "ref",
          "head_sha"
        ],
        "title": "IngestPrRequest",
        "type": "object"
      },
      "IngestPrResponse": {
        "properties": {
          "change_id": {
            "format": "uuid",
            "title": "Change Id",
            "type": "string"
          },
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "diff_truncated": {
            "title": "Diff Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "change_id",
          "diff_truncated",
          "created"
        ],
        "title": "IngestPrResponse",
        "type": "object"
      },
...
      "RetryReviewResponse": {
        "properties": {
          "change_id": {
            "format": "uuid",
            "title": "Change Id",
            "type": "string"
          },
          "run": {
            "title": "Run",
            "type": "integer"
          }
        },
        "required": [
          "change_id",
          "run"
        ],
        "title": "RetryReviewResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Duelo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/changes/{change_id}": {
      "get": {
        "operationId": "get_change_changes__change_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Change Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Change",
        "tags": [
          "changes"
        ]
      }
    },
    "/changes/{change_id}/events": {
      "get": {
        "description": "L\u00ednea de tiempo del change: `change.created`, `review.completed` y `review.failed`.",
        "operationId": "get_change_events_changes__change_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Change Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChangeEventResponse"
                  },
                  "title": "Response Get Change Events Changes  Change Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Change Events",
        "tags": [
          "changes"
        ]
      }
    },
    "/changes/{change_id}/retry": {
      "post": {
        "description": "Nueva ejecuci\u00f3n de la review de un change cuya ejecuci\u00f3n actual termin\u00f3 con fallos.",
        "operationId": "retry_review_changes__change_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Change Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-ingest-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Token"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Retry Review",
        "tags": [
          "changes"
        ]
      }
    },
...
    "/ingest/commit": {
      "post": {
        "operationId": "ingest_commit_ingest_commit_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-ingest-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestCommitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestCommitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Ingest Commit",
        "tags": [
          "ingest"
        ]
      }
    },
    "/ingest/pr": {
      "post": {
        "operationId": "ingest_pr_ingest_pr_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-ingest-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestPrRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestPrResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Ingest Pr",
        "tags": [
          "ingest"
        ]
      }
    },
...
    "/projects/{slug}/changes": {
      "get": {
        "operationId": "list_changes_projects__slug__changes_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ChangeKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ChangeReviewStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "pattern": "^[^\\x00]*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Changes",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}